updates to tests
